Disable Codex docs build for now to avoid OOM issue
The build process for the Codex docs runs out of memory. The docs are
not that important for patchdemo, so rather than trying to increase the
amount of available memory, just disable the docs build for now.

Also copy files for the design tokens package, and update the list of
files that is copied for the icons package.

This is a bit messy, but hopefully we can make all of this code obsolete
in the future by implementing T314507.