Run processes in parallel
Use process->start instead of process->run to allow a
group of processes to run in parallel.

Depending on the system bottleneck and number of wikis
selected, this can make setting up a wiki much faster.

When checking out repos, some are dependent on others,
so core and VE-MW are run synchronously.

Locally, with all repos update to date, a WMF wiki
goes from taking 83s to 37s.

In production, WMF wikis currently take about 132s when
all repos are up to date.