Fix installation failure on wmf.X versions without ORES
When picking a version like wmf/1.43.0-wmf.4, the extensions/ORES/
directory will always exist, because it's created by a submodule in
mediawiki/core, but it will be empty if the ORES extension is not
enabled, causing this step to fail.

Check the existence of the maintenance script instead, like we do for
AntiSpoof earlier in this file.