Copy all Codex files, don't try to list them all
Codex keeps adding new files do the dist directories of its packages, so
the lists and globs we have in our script keep getting out of date. To
avoid this, just copy everything in the dist directory to resources/lib,
even if that includes some files that MediaWiki doesn't use.