Enable temp accounts by default
As part of the general effort to acquiant the technical and product
community with temporary accounts, it would be helpful to set the
default on Patch Demo wikis to enable temporary accounts, ahead of the
work to enable temporary accounts across WMF wikis in 2024 and 2025.

By enabling by default, we increase the likelihood of surfacing edge
case technical and UX issues prior to deployment to production wikis.

Change-Id: I61be119b87016eed93463fad0f16e750b63d1cc5