Adds sidecar containers for log collection
patchdemo runs apache2 which writes logs to /var/log/apache2. In order
to collect these logs through normal kubernetes log collection, we made
/var/log/apache2 into a shared mount. An init container is required to
set the permissions of the mount (which default to root ownership). The
three sidecar containers tail the logs to stdout which is picked up by
kubernetes and can be seen using the normal `kubectl log` command.

Bug: T370078

Closes T370078