Catalyst backend: support all extensions and skins repositories
Including the atypical `VisualEditor/VisualEditor` and
`mediawiki/services/parsoid` repos.

`design/codex` and `oojs/ui` are still pending.

Also fixed a bug were we were sending the Catalyst API the full
list of allowed repositories instead of only the ones selected by the
user. This was probably not detected previously because when
the preset `custom` is selected the list of allowed repos is the
same as the selected repos. For other presets those two lists differ

Bug: T373604