Unlock session after writing
Leaving the session locked doesn't allow users to navigate anywhere else until the script stops running.

This affects the user during the catalyst wiki creation log stream. We can also check the connection and abort the script, but this is slower since the check must wait for the next chunk in the log stream to arrive. In addition, when creating wikis using the catalyst backend, we want the script to continue running to update the wiki status and TTC information in the database even if the user navigates away from the log stream.

* Adds session_write_close() to the end of setup in Authentication.php
* Moves setting csrf token to inside setup where session is locked

Bug: T385073