README: fix image URLs
The URLs to the images at the top of `README.md`
currently point to the `master` branch, when this
repository uses the `main` branch.

Updated the image URLs to use the `main` branch
instead of the `master` branch.

---

(first GitLab merge request - hopefully I haven't messed anything up too badly!)